sem11/test_api/requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.1.1
djangorestframework==3.8.2
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.1.1
django-guardian==1.4.9
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
